[PSEC-1106] Move composite test clients in a composite package
