+++ tests/clients/composite/test_aws_central_logging_client.py
