+++ tests/clients/composite/test_aws_cloudtrail_client.py
