+++ tests/clients/composite/test_aws_vpc_client.py
